% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATI-package.R
\docType{package}
\name{ATI-package}
\alias{ATI}
\alias{ATI-package}
\title{ATI: Algorithmic trading and investment}
\description{
A package for teaching a graduate course in algorithmic trading and investment. The course is based on the emerging field of financial machine learning.
}
\author{
\strong{Maintainer}: Barry Quinn \email{b.quinn@qub.ac.uk} (\href{https://orcid.org/0000-0002-8637-9060}{ORCID})

}
\keyword{internal}
